default to echo package
remove tasks. copy some contents as example in comments
