+++ common/tools/dev-tool/src/commands/admin/list/packages.ts
@@ -20,6 +20,43 @@ export const commandInfo = makeCommandInfo("packages", "list packages defined in
     kind: "string",
     shortName: "s",
   },
+
+  // use --task option to provide code file to be excuted for each package. For example,
+  // the following code adds missing "tsx" dev dependency whereever it is used.
+  /*
+import * as path from "node:path";
+import { writeFile } from "node:fs/promises";
+import { resolveProject } from "../../../../util/resolveProject";
+import { RushJsonProject } from "../packages";
+
+export default async function IdentifyPackage(
+  project: RushJsonProject,
+  _paths: boolean,
+  _cwd: string,
+  root: string,
+) {
+  const { projectFolder, packageName } = project;
+  const fullProjectPath = path.join(root, projectFolder);
+  const { packageJson } = await resolveProject(fullProjectPath);
+
+  const unitTestNodeScript = packageJson.scripts["unit-test:node"];
+  const integrationTestNodeScript = packageJson.scripts["integration-test:node"];
+
+  if (unitTestNodeScript.includes("dev-tool run test:node-js-input") ||
+    integrationTestNodeScript.includes("dev-tool run test:node-js-input") ||
+    unitTestNodeScript.includes("dev-tool run test:node-tsx-ts") ||
+    integrationTestNodeScript.includes("dev-tool run test:node-tsx-ts")) {
+    if (!packageJson.devDependencies["tsx"]) {
+      console.log(`    updating ${packageName} ${fullProjectPath} testing scripts`);
+      packageJson.devDependencies["tsx"] = "^4.7.1";
+      await writeFile(
+        path.join(fullProjectPath, "package.json"),
+        JSON.stringify(packageJson, undefined, 2) + "\n",
+      );
+    }
+  }
+}
+*/
   task: {
     description: "a .js/.ts file default-exporting a function to run on each listed package",
     kind: "string",
@@ -66,26 +103,28 @@ export async function getProjects(service?: string): Promise<RushJsonProject[]> 
     : rushJson.projects;
 }
 
+async function echoPackage(project: RushJsonProject, paths: boolean, cwd: string, root: string) {
+  if (paths) {
+    console.log(path.relative(cwd, path.resolve(root, project.projectFolder)) || ".");
+  } else {
+    console.log(project.packageName);
+  }
+}
+
 export default leafCommand(commandInfo, async ({ paths, service, task }) => {
   const cwd = process.cwd();
   const root = await resolveRoot();
 
   const projects = await getProjects(service);
 
+  let func: any;
   if (task) {
-    const imported = await import(task);
-    console.dir(imported.default);
-    for (const project of projects) {
-      await imported.default(project, paths, cwd, root);
-    }
+    func = (await import(task)).default;
   } else {
-    for (const project of projects) {
-      if (paths) {
-        console.log(path.relative(cwd, path.resolve(root, project.projectFolder)) || ".");
-      } else {
-        console.log(project.packageName);
-      }
-    }
+    func = echoPackage;
+  }
+  for (const project of projects) {
+    await func(project, paths, cwd, root);
   }
 
   return true;
