--- common/tools/dev-tool/src/commands/admin/list/tasks/echo-packages.js
@@ -1,10 +1,0 @@
-const path = require("path");
-
-export default async function EchoPackage(project, paths, cwd, root) {
-  console.dir({
-    project,
-    paths,
-    cwd,
-    root,
-  });
-}
